% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{f_test_RSS}
\alias{f_test_RSS}
\title{Compute F-statistic and p-value for linear restrictions (model comparison)}
\usage{
f_test_RSS(y, X, C, d)
}
\arguments{
\item{y}{Response vector}

\item{X}{Design matrix}

\item{C}{Constraint matrix}

\item{d}{Constraint vector}
}
\value{
List with F_statistic and p_value
}
\description{
Compute F-statistic and p-value for linear restrictions (model comparison)
}
\examples{
y <- c(1,2,3,4)
X <- cbind(1, c(1,2,3,4))
C <- matrix(c(0,1), nrow=1)
d <- 0
# compute_F_stat(y, X, C, d)
}
