% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{find_beta}
\alias{find_beta}
\title{Compute the OLS estimate of regression coefficients}
\usage{
find_beta(y, X)
}
\arguments{
\item{y}{Response vector}

\item{X}{Design matrix}
}
\value{
Numeric vector of OLS estimates
}
\description{
Compute the OLS estimate of regression coefficients
}
\examples{
y <- c(1,2,3,4)
X <- cbind(1, c(1,2,3,4))
find_beta(y, X)
}
