% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs.R
\name{find_betaH}
\alias{find_betaH}
\title{Compute constrained OLS estimate subject to linear constraints}
\usage{
find_betaH(y, X, C, d)
}
\arguments{
\item{y}{Response vector}

\item{X}{Design matrix}

\item{C}{Constraint matrix}

\item{d}{Constraint vector}
}
\value{
Numeric vector of constrained OLS estimates
}
\description{
Compute constrained OLS estimate subject to linear constraints
}
\examples{
y <- c(1,2,3,4)
X <- cbind(1, c(1,2,3,4))
C <- matrix(c(0,1), nrow=1)
d <- 0
# find_betaH(y, X, C, d)
}
